+++ rs/anda_db/tests/coverage_public_api.rs
@@ -5,8 +5,8 @@ use anda_db::{
     index::{BM25, BTree, Hnsw, HnswConfig, default_tokenizer},
     query::{Filter, Query, RRFReranker, RangeQuery, Search},
     schema::{
-        AndaDBSchema, ByteBufB64, Document, FieldEntry, FieldType, Fv, Schema,
-        SchemaError, Vector, bf16,
+        AndaDBSchema, ByteBufB64, Document, FieldEntry, FieldType, Fv, Schema, SchemaError, Vector,
+        bf16,
     },
     storage::{PutMode, Storage, StorageConfig, StorageStats},
 };